統合途中

+++ master/DebriEmitter.cpp
@@ -10,7 +10,7 @@ DebriEmitter::~DebriEmitter()
 	debris.clear();
 }
 
-void DebriEmitter::Emit(const Vector2Int& leftTop, const Vector2Int& rightBottom, const int num, 
+void DebriEmitter::Emit(const Vector2Int& leftTop, const Vector2Int& rightBottom, const size_t num,
 					const Debri::Color color[2], const int graph)
 {
 	debris.clear();
@@ -20,7 +20,7 @@ void DebriEmitter::Emit(const Vector2Int& leftTop, const Vector2Int& rightBottom
 		Debri debri;
 		Vector2Int pos = { GetRand(leftTop.x, rightBottom.x), GetRand(leftTop.y, rightBottom.y) };
 		float scale = GetRandF(0.5f, 1.5f);
-		float rota = GetRandF(0, 3.141592 * 2.0f);
+		float rota = GetRandF(0, 3.141592f * 2.0f);
 		Vector2Int speed = { GetRand(-5, 5), GetRand(-10, 0) };
 		int ary = GetRand(0, 1);
 		Debri::Color c = color[ary];
@@ -33,7 +33,7 @@ void DebriEmitter::Emit(const Vector2Int& leftTop, const Vector2Int& rightBottom
 
 void DebriEmitter::Update()
 {
-	int count = 0;
+	size_t count = 0;
 	for (size_t i = 0; i < debris.size(); i++)
 	{
 		if (debris[i].isDeath) count++;
