+++ master/Ease.cpp
@@ -7,11 +7,11 @@ float lerp(const float a, const float b, const float t)
 }
 float EaseIn(const float start, const float end, const float time, const float power)
 {
-	return lerp(start, end, pow(time, power));
+	return lerp(start, end, powf(time, power));
 }
 float EaseOut(const float start, const float end, const float time, const float power)
 {
-	return lerp(start, end, 1 - pow(1 - time, power));
+	return lerp(start, end, 1 - powf(1 - time, power));
 }
 
 Ease::Ease() :
