+++ master/Random.cpp
@@ -3,7 +3,7 @@
 #include <math.h>
 #include <time.h>
 
-void InitRand() { srand(time(NULL)); }
+void InitRand() { srand((int)time(NULL)); }
 
 int GetRand(const int start, const int end)
 {
@@ -22,5 +22,5 @@ float GetRandF(const float start, const float end)
 	if (div < 0.0f) div -= 1.0f;
 	if (div == 0.0f) return end;
 
-	return (fmodf(rand(), div) + start);
+	return (fmodf((float)rand(), div) + start);
 }
